pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cyberguard-mlops"
version = "0.1.0"
description = "Cybersecurity-themed MLOps intrusion detection pipeline with FastAPI serving, MLflow, DVC, Airflow, and monitoring."
requires-python = ">=3.11,<3.12"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.8",
  "numpy>=1.26",
  "pandas>=2.2",
  "scikit-learn>=1.5",
  "joblib>=1.4",
  "mlflow>=3.11,<4.0",
  "dvc>=3.59",
  "great-expectations>=1.3",
  "evidently>=0.6",
  "prometheus-client>=0.20",
  "python-multipart>=0.0.9",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "pytest-cov>=5.0",
  "black>=24.8",
  "isort>=5.13",
  "ruff>=0.6",
  "mypy>=1.11",
  "pre-commit>=3.8",
]
airflow = [
  "apache-airflow>=3.0,<4.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
follow_imports = "skip"
disallow_subclassing_any = false
disallow_untyped_decorators = false
warn_unused_ignores = false

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
